rename section files
