+++ src/components/sections/Hero.section.styles.ts
