+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
-import { SectionAbout } from '../components/sections/SectionAbout';
-import { SectionArticles } from '../components/sections/SectionArticles';
-import { SectionHero } from '../components/sections/SectionHero';
-import { SectionProjects } from '../components/sections/SectionProjects';
+import { SectionAbout } from '../components/sections/About.section';
+import { SectionArticles } from '../components/sections/Articles.section';
+import { SectionHero } from '../components/sections/Hero.section';
+import { SectionProjects } from '../components/sections/Projects.section';
 import { ToolBox } from '../components/toolbox/ToolBox';
 import { ButtonPrimary } from '../components/ui-system/buttons/ButtonPrimary';
 import { Column } from '../components/ui-system/container/Column';
